microblog/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
Pillow
channels